refactor(suite-native): rename banner-flags package to banners
Widen the package's naming to reflect its growing responsibility beyond
dismissal-flag state, ahead of adding a shared Banner UI component.
