+++ suite-native/banners/package.json
@@ -1,5 +1,5 @@
 {
-    "name": "@suite-native/banner-flags",
+    "name": "@suite-native/banners",
     "version": "1.0.0",
     "private": true,
     "license": "See LICENSE.md in repo root",
