+++ suite-native/module-device-onboarding/tsconfig.json
@@ -22,7 +22,7 @@
         { "path": "../alerts" },
         { "path": "../analytics" },
         { "path": "../atoms" },
-        { "path": "../banner-flags" },
+        { "path": "../banners" },
         { "path": "../config" },
         { "path": "../device" },
         { "path": "../device-authorization" },
