+++ suite-native/module-home/src/screens/HomeScreen/components/OnboardingFeedbackBanner.tsx
@@ -5,7 +5,7 @@ import { events } from '@suite-common/analytics';
 import { useServices } from '@suite-common/dependency-injection';
 import { selectNativeAnalyticsDep } from '@suite-native/analytics';
 import { Box, Button, HStack, IconButton, Text, VStack } from '@suite-native/atoms';
-import { setIsOnboardingFeedbackBannerEnabled } from '@suite-native/banner-flags';
+import { setIsOnboardingFeedbackBannerEnabled } from '@suite-native/banners';
 import { Translation } from '@suite-native/intl';
 import { useOpenLink } from '@suite-native/link';
 import { prepareNativeStyle, useNativeStyles } from '@trezor/styles-native';
