+++ suite-native/module-settings/package.json
@@ -35,7 +35,7 @@
         "@suite-native/alerts": "workspace:*",
         "@suite-native/analytics": "workspace:*",
         "@suite-native/atoms": "workspace:*",
-        "@suite-native/banner-flags": "workspace:*",
+        "@suite-native/banners": "workspace:*",
         "@suite-native/biometrics": "workspace:*",
         "@suite-native/coin-enabling": "workspace:*",
         "@suite-native/config": "workspace:*",
