+++ suite-native/module-settings/tsconfig.json
@@ -45,7 +45,7 @@
         { "path": "../alerts" },
         { "path": "../analytics" },
         { "path": "../atoms" },
-        { "path": "../banner-flags" },
+        { "path": "../banners" },
         { "path": "../biometrics" },
         { "path": "../coin-enabling" },
         { "path": "../config" },
