+++ suite-native/state/package.json
@@ -39,7 +39,7 @@
         "@suite-common/wallet-core": "workspace:*",
         "@suite-common/walletconnect": "workspace:*",
         "@suite-native/analytics": "workspace:*",
-        "@suite-native/banner-flags": "workspace:*",
+        "@suite-native/banners": "workspace:*",
         "@suite-native/biometrics": "workspace:*",
         "@suite-native/blockchain": "workspace:*",
         "@suite-native/bluetooth": "workspace:*",
